pm: support verified dexopt for allowlisted packages
Packages that are allow-listed in a system config will have their PackageState marked as isVerifiedDexopt.
When ART dexopts a package it will look at this flag and then compile the APK in a PVM instead of running dex2oat directly on the Android host.

Bug:439646165
Bug:423718710
API-Coverage-Bug: 445170833
Test: atest SystemConfigTest, manual testing via pm dump-package
Flag: android.content.pm.verified_dexopt
Change-Id: I55435387b7cab721b082ca4a8aff12ac487a7e97