Fix: Resolve crash from delayed main thread tasks
A delayed message posted to the main thread can cause a crash if it's processed after a test has completed. This change adds a cleanup step to clear all pending tasks on the main thread's handler in the test tear-down. This prevents test pollution and ensures a clean, isolated environment for each test run.

Flag: EXEMPT fix flaky test
Bug: 443688194
Test: atest HearingDevicesDialogDelegateTest KeyGestureDialogInteractorTest
Change-Id: I57373af02d57d0e171b52c923d62a6a66c9232e3