Use pending transaction for window animation of closing app
In case the dialog has shown on the visible activity but the
reparent of animation leash is not applied yet by transition.

The original case of using sync transaction is only needed for
a visible activity. Because if the activity is opending with
a window level animation, the reparent of leash can be applied
with transition to avoid the window from being visible before
the window level animation starts.

Bug: 443626141
Flag: EXEMPT bugfix
Test: An activity shows a dialog and finishes itself
      at the same time. The dialog won't flicker a frame
      at top position.
Change-Id: Ie1ee1bd4b534d92e7a1fcac7d22273b6f5f91ecc