Fix unit test crash on eng builds due to Log.wtf
A unit test was crashing on engineering builds because a call to
flushAll() was executing a pending UncertainTracker message. This
intentionally triggers a Log.wtf() crash, which occurs after SWC posts
the tracker to the mainExecute queue.
The fix is to remove the pending UncertainTracker message from the
queue before flushAll() runs. The test is also updated to verify that
removeCallbacks() is correctly called to verify the transactionCommit()
will remove the tracker.
Also shorter the UNCERTAIN_TRANSITION_TIMEOUT_MS since the
startTransaction should normally applied right after the transition
is handled.

Bug: 431189806
Flag: EXEMPT bugfix
Test: atest StartingWindowControllerTests
Change-Id: I3295257bee5f9f50f3456ef6db08f714468f5bf8