a11y: Ignore the unittest causing crash
We have two Magnification Delegate dialogs. For the old one, we don't
set dismissal for it; for the new one, we set dismissal for it to
disable shortcut. The new one is behind a flag, which we enable it by
default recently. Thus, the crashing unittest is creating the new Mag
dialog which will crash during `tearDown()`. We already have one CL
ag/35510953 to update the test fixture for fixing the dismissal crash.

This CL will ignore the crashing unittest which is tempopry fix to let
it not affect others in main branch.

Bug: b/443556122
Flag: com.android.hardware.input.enable_magnify_magnification_key_gesture_dialog
Test: atest KeyGestureDialogStartableTest
Change-Id: Iec4ecb84d9895debaea8d338f3afca0376624e9e