Implement the new edit mode top app bar
This changes the top bar to expand/collapse based on the scrolling of the page. We display the page's subtitle in the expanded version instead of having it as a header of the grid. This also changes the colors of the undo button, a follow-up CL will add the remove button next to it.

Some refactoring was done for this change, but the current behavior and UI is unchanged.

Flag: com.android.systemui.qs_edit_mode_v2
Bug: 438943936
Test: manually - edit tiles in edit mode
Test: EditModeTest
Change-Id: If4dc2ee99e24be6230ecad7cc2b6c1bd3152d667