VirtualInputDeviceController: Use input device phys to identify device
Use the "phys" identifier to identify a virtual input device upon
creation, instead of using the name. The input device name is part of
the exposed API surfaces, so we should not have artificial requirements
in such surfaces, like requiring the name to be unique. Since the phys
is uniquely generated by the system, it should be used to uniquely
identify the device instead.

Bug: 443074681
Bug: 443736238
Test: atest InputTests, manual with hellocomputercontrol app
Flag: EXEMPT bug fix
Change-Id: I97fe38ae251bc686561924e1a3f888bcae9da391