[1/n] Improve restart button heuristic
At the moment the restart button is hidden if the top activity
is transparent without considering if it's the only one in the
stack. The initial idea was to hide the button only in case
of permission dialogs or, in general, in case the transparent
activity had opaque activities below.

Using TransparentPolicy allows to hide the restart button only
when the transparent activity has opaque activities below. In case
of a completely transparent activity with no opaque activities
below, the button should then be visible if the activity is
in size compat mode.

Flag: com.android.window.flags.app_compat_ui_restart_on_full_transparency
Bug: 444422899
Test: atest WmTests:AppCompatUtilsTest

Change-Id: Iace0b5f024286cfc8d6a48efb9592f3bf0e72bec