Apply negative margin to the correct view
In order for all the QS views to extend to the bottom of the screen, we
applied a negative margin to counteract the padding of
NotificationsQuickSettingsContainer. However, this was applied to the
wrong view. Instead of applying it to the top level qs_frame, we applied
it to quick_settings_container (the ComposeView). This meant that the
ComposeView was larger than its parent (and grandparent) and therefore
touches were not delivered properly.

With this, we apply the negative margin to qs_frame, which is the direct
child of NotificationsQuickSettingsContainer. Then all the views extend
to the bottom and touches are not intercepted.

Test: manual, rotate unfolded device and touch footer actions
Fixes: 444310993
Flag: EXEMPT BUG_FIX

Change-Id: I172cfd0041a5c4138f61a247492e9190110c4123