Add test to verify letterbox insets are cleared when hidden
Adds a unit test for the `ShellLetterboxPolicyState.hide()` method. This test ensures that when the letterbox is hidden, all its associated bounds are cleared.

This prevents stale letterbox inset data from being used in other parts of the system, such as for task snapshots, after the letterbox is no longer visible.

Test: ATP tests passed http://go/forrest-run/L55000030017916714
Bug: 431235865
Flag: TEST_ONLY

Change-Id: I63828bfe1fbc6303b405c94c8d7ed3b0a03be8f6