Add tests for debugging feature restriction checks
Adds unit tests for the `isSettingRestrictedForUser` method to verify the logic for the `DISALLOW_DEBUGGING_FEATURES` restriction.

The new tests cover scenarios where the restriction is applied by the system user, by a separate device owner, or not at all. They also verify edge cases, such as when the device owner is the system user and when a user attempts to disable the setting (which is always allowed).

These tests indirectly cover the new private helper methods `getDeviceOwnerUserId` and `deviceOwnerOrSystemUserHasRestriction`, ensuring the global restriction check works as intended.

Test: With ag/35606994 reverted, failing (http://ab/I39300010446086097); with ag/35606994 in the codebase, passing (http://ab/I53100010446400053), AND ATP tests passed http://go/forrest-run/L86300030017888689
Bug: 431235865
Flag: TEST_ONLY

Change-Id: I3c9a88fe7039d401e918073d28dfc1202169184e