Use a background CoroutineContext that is backed by a HandlerThread for eagerly creating display instances
Problem:
The @Background CoroutineScope is backed by a fixed thread pool. This means the thread associated with this scope lacks an associated Looper object. Consequently, when Dagger attempts to provide dependencies that rely on a Looper, the dependency injection fails because the required Looper is null. We need to ensure the @Background scope/dispatcher is initialized with a proper HandlerThread to fix this. The Dagger provided @Background Executor is a good choice because it is backed by a HandlerThread.

Bug: 443801583
Flag: com.android.window.flags.enable_connected_displays_wallpaper_presentations
Test: atest SystemUITests:MultiDisplayCursorPositionRepositoryTest
Test: atest SystemUITests:PerDisplayInstanceRepositoryImplTest
Change-Id: I0efdfcfda46378c9f7fb5e6559dd7373b88672ec