Recover from broken bubble gesture monitor state
If the bubble swipe up listener is active while collapsed, stop
monitoring gestures. This recovers from the state where swiping to
go home or to overview doesn't work.

Note that this isn't ideal because we still don't have reliable
repro steps, but with this change, the first time the user swipes up
will not work as expected but will recover and work as expected the
second time, instead of staying stuck in the broken state.

Bug: 441574611
Flag: EXEMPT BUGFIX
Test: manual - verify swipe up logic works as expected
Test: manual - made some changes in the code to force enter a broken
       gesture monitor state and verified that swiping up the second
       time works.
Change-Id: Ib528dcb4a37b0fb9cdbaa338e1baed75c70d038b