Tighten EnforcingAdmin instantiation: flag
Currently there are 5 static factory methods, 4 constructors, and an XML
deserializer, each of which contains precondition checks. This makes it
hard to understand what are the allowed states for EnforcingAdmin
instances. Moreover, when creating EnforcingAdmins from XML or parcels,
this class allows pretty much any combination of authorities, even those
that cannot arise when a policy is set in DPMS.

CLs in this sequence aim to narrow this down, such that all
instantiation of EnforcingAdmins goes through one of the static factory
methods which enforces all the preconditions and sets all the default
values.

This will unblock further refactorings, in particular, extracting
AdminKey, such that policies are keyed on immutable attributes of admin
entity, not dynamic parts of package state (e.g. roles).

Bug: 335663055
Test: atest FrameworksServicesTests:com.android.server.devicepolicy.EnforcingAdminTest
Flag: android.app.admin.flags.tighten_admin_instantiation
Change-Id: If06bacd4afc6127ef497a8b7b14a2c6950713a5b