Tighten EnforcingAdmin instantiation from binder
createEnforcingAdmin is used to create internal EnforcingAdmins from
parcelable a.a.a.EnforcingAdmin and it is only invoked from
getPolicySizeForAdmin TestApi. This CL changes it to use one of 4 static
factory methods to create a well-formed admin and prevents creation of
admins with nonsensical combinations of roles.
Previously roles were copied from the incoming a.a.a.EnforcingAdmin, now
they are not - they will be fetched from role manager if necessary.

Bug: 335663055
Test: atest FrameworksServicesTests:com.android.server.devicepolicy.EnforcingAdminTest
Flag: android.app.admin.flags.tighten_admin_instantiation
Change-Id: I6527058d98489a98b062d5541402b4170e0ae7f2