Clean up AnrTimer options
Android always uses native AnrTimers, with the exception of non-linux
development environments.  AnrTimer automatically falls back to a
MessageQueue implementation in non-linux environments. The Args option
to select non-native timers is removed.  Non-native timers can be
selected by a new Injector option; this is only used in unit tests to
verify non-native timer functionality.

The Args testMode option is moved into the Injector.  It is another
option that is only relevant to test code.

The AnrTimer serviceEnabled method is removed.  It is unnecessary now
that Android always uses native timers.

AnrTimerTest is slightly refactored to account for the above changes
and to improve readability.

Flag: EXEMPT PURE_REFACTOR
Bug: 445189898
Test: atest
 * FrameworksServicesTests:AnrTimerTest
Change-Id: I4c0e92b9c3ea88c5eff27f5e3115fba95a526c8c