Improve clarity of skipping return to drag start animation
In the block handling the case where destinationBounds and
dragStartBounds are the same, the conditional check to skip the
animation was comparing currentDragBounds to dragStartBounds, due to a
recent change Ib7393d109c12686cab6440b58ccfe91049f18a94.
While logically equivalent in this scope, changing it back to compare
against destinationBounds `if (destinationBounds != currentDragBounds)`
makes the intent clearer and align with the inline comment.

Flag: EXEMPT PURE_REFACTOR
Bug: 442613884
Test: atest DesktopTasksControllerTest
Change-Id: Id15996acbdee788710bb78fbe8cff03ffd0f0b6f