[Flexiglass] Remove unnecessary adjustment of HUN drawBounds
The drawBounds communicated from the SnoozeableHeadsUpNotificationSpace
to the NSSL are wrongly modifierd by headsUpInset. This is a leftover
from a previous bugfix, but by now reordering the modifiers made it
redundant:
 - apply the offset first
 - add the onGloballyPositioned callback later

The extra headsUpInset added to the height of the drawBounds is making
the drawBounds Rect too tall now. This CL removes it.

Fixes: 445618359
Test: check debug bounds of HUNs over GONE and/or QSShadeOverlay
Test: check whether the bottom area of a HUN is clickable
Flag: com.android.systemui.scene_container
Change-Id: I9e8963163872346db324b19c43ffd435dfd55097