Allow invisible PiP task to close with request
Generally when a container is not visible (e.g. when display is off),
we are currently not letting any changes to prpogate thru transitions.
This includes existence changes, like a container being removed.

With PiP this can happen when an app directly finishes the PiP activity
while the screen is off (app can listen to onStop being received).
This would result in the PiP task being removed, however since PiP2
fully relies on Transitions, Shell would not receive info about the
closing task, or be aware of it at least through the transition request.

This in turn, leaves PiP state stale on the Shell-side, as PiP thinks
the PiP task is still present and might block enter PiP or enter
improperly on the follow-up sessions.

Bug: 444345763
Flag: com.android.window.flags.transit_invisible_existence_change
Test: enter PiP from Netflix then turn the display off
Change-Id: I08fd4f50f63437e008834ddaa50a8ae11872fd7a