[Dual Shade] Extract more ShadeMode logic into a configuration resource.
First, this removes the `config_disableSplitShade` overlay. We no longer
want to have desktop-specific behavior for Dual Shade. Instead, we use
properties of the current shade window and configuration to determine
the shade mode.

This CL also introduces config values for whether Dual Shade should be
turned on by default, and whether the user's chosen setting should be
overriden (for example, we do not intend to show the setting on large
screens). The latter will now also used to inform the Dual Shade setting
visibility, as well as the visibility of the shortcut to said setting in
Quick Settings Edit Mode.

Fix: 442479130
Bug: 444424242
Test: Unit tests still pass.
Test: Verified manually by enabling and disabling Dual Shade on both a
 compact and large screen device.
Flag: com.android.systemui.scene_container
Change-Id: Ib3237bc21b530730126d764ba2659ca9028a7ccc