Call Profiling on Application Crash
Notify ProfilingService when an app encounters a crash. Block AMS from
killing the app until ProfilingService is done, using a timeout
provided by ProfilingServiceHelper. ProfilingService will be "done"
when either the profiling finishes, or when it determines not to run
profiling.

Current use of this is limited to collecting a java heap dump on an
OOM crash with a delay timeout of up to 5 seconds.

Test: manually trigger an OOM for registered process, observe trigger
      and processing. Trigger for unregistered process, observe no
      delay with killing app at await.
Test: new tests in topic'd CL
Bug: 441723167
Flag: android.os.profiling.profiling_trigger_oom
Change-Id: Ic80f0b7da29b12be54da1ae76ee474399426d7b9