Add displayId to WindowManagerPolicy#isScreenOn().
This allows certain flows to continue to work even if the default
display is off. For example, when taking an Assist screenshot for
Contextual Search, DisplayContent checks if the screen is on. This
previously only checked the default display, but now instead checks
the display where the search is invoked.

Fix: 441675960
Test: Connect phone to external display, wait for primary display to
time out, invoke Contextual Search on the external display.
Test: atest DisplayPolicyTests WindowManagerServiceTests
Flag: EXEMPT BUGFIX

Change-Id: I21ae35333551e170127be546d1cc8ca09bad4448