am: Use ActiveUidsInternal in OomAdjuster
This CL refactors OomAdjuster to use the `ActiveUidsInternal`
instead of the `ActiveUids` class from ActivityManagerService.
All usages of `ActiveUids` and `UidRecord` are replaced with their
`Internal` counterparts within `OomAdjuster` and related classes.

Since other parts of ActivityManagerService have not yet been updated,
this CL introduces temporary casts from `UidRecordInternal` to
`UidRecord` when calling into AMS. These are marked with a TODO and will
be removed in a subsequent change.

Bug: 446555028
Test: m services.core
Test: atest MockingOomAdjusterTests OomAdjusterTests
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I5c75d2296b6d24659b420b177c1289555e1e4f0d