Remove redundant checks from sleep-for-shutdown.
Shutdown logic is doing multiple sleep checks. This
is unnecessary work.

On its own, this isn't really important; however, this
is part of a larger refactor to separate callpaths
with hierarchy impact from callpaths without.

This also exposed that `allowDelay` plumbing isn't
super necessary and the leaf-level `putTasksToSleep`
can be split into delay and non-delay and then we
don't have to plumb the delayability around.

Bug: 423701190
Test: refactor, so existing tests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I1eec0ac8509ab9833010e6435e9c8088e8dc4d8a