Fix crash that can happen with nested STLs
This CL fixes a crash that can happen in STL when the wrong transition
of an ancestor STL was incorrectly picked for the transition of an
element. On top of checking that the element is shared or has a
transformation on one of the transition scenes, we now also check that
the element is currently in that scene.

Bug: 442640840
Test: NestedSceneTransitionLayoutTest
Flag: EXEMPT BUGFIX
Change-Id: Ic8e3340937209d738cfcb3b28bdb23c3bde6aace