Fix keyguard presentation stuck on default display after the shade moves (race)
Every time the device is locked the shade window goes back to the default display.
However, if the code that shows the presentation is triggered early enough, it could be adding the presentation to the default display (as its logic is to add the presentation to all displays without the shade window, that already has its own lockscreen)

After this change the shade move signal is taken into consideration to update where the presentation is.

Also, to avoid flickers (e.g. first adding the presentation to the default display, and once the shade move finished removing it), the shade display "pending display id" signal is used instead of the "display id" one. This gets updated even before the reparenting is finished. If the reparenting fails, it will go back to the correct display anyway.

Fix: 447148049
Test: DisplayWallpaperPresentationInteractorTest
Flag: com.android.window.flags.enable_connected_displays_wallpaper_presentations
Change-Id: I848148a070c07f463a9771d9427c43d410d46ec5