Move widget flag to android.app.Flags
I just learned that gantry does allow metadata to change on a flag (go/trunk-stable-bug-fix#switching) and the added refactor flag relies on a separate flag but changes are related. Since I'm in a state where the documentation recommends I re-name the flags anyway, I might as well combine them to simplify the roll-out for the change and the refactor.

Flag: android.app.notification_transparent_badge_ring
Test: N/A, flag change only
Bug: 442592454, 409346132
Change-Id: I2dcd3dd49be4ba722a093d3d8d129e892edf9659