BouncyBall: Switching to FrameMetrics for in-app detection
Using Choreographer.FrameCallback.doFrame(long frameTimeNanos)
and diffing timestamps is not a reliable way to detect frame
drops within an app.

Much closer is android.view.FrameMetrics, so we switch to that.
However, since this requires an additional callback, and in-app
detection still isn't as completely reliable as tracing, we
turn this off by default.

Note that this logic not only catches missed_app_frames, but
also missed_sf_frames.  That's a problem with how our trace
metrics logic works right now (it ignores missed_sf_frames).
But per b/447197023, we're going to start failing on
missed_sf_frames as well, so this will become consistent.

Test: atest --include-preview-tests GtsPerformanceHostTestCases
Bug: 437906487
Flag: EXEMPT TEST_ONLY
Change-Id: Icb0b35c36f76660fcaf224ff6359562275cefa05