Fix dragging bubble bar expanded view to dismiss
Previously the dismiss target was consuming all of the touch events,
and DropTargetManager wasn't able to track drag zones. This change
ensures that DropTargetManager is always notified.

Also at the end of the drag we now notify launcher with the final
bubble bar location when dragging to dismiss.

Fixes: 445762164
Flag: com.android.wm.shell.enable_bubble_bar
Test: atest BubbleBarLayerViewTest
Test: manual
       - have 2 bubbles in bubble bar
       - drag expanded view to the other side and then to dismiss
       - observe bubble bar and expanded view animate back to initial
         location
       - drag the expanded view again to the other side and dismiss
       - verify bubble bar was dismissed
       - create new bubble and verify expanded view and bubble bar
         are positioned in the initial location
Change-Id: Iaaf3b78fa87144a296d1ec9e356e30dddba57665