am: Move post-oom-adj logic to a new callback
This CL refactors the logic that runs after an OOM adjustment pass is
complete. Previously, OomAdjuster was responsible for tasks like
scheduling the destruction of activities for "always-finish" and
triggering process stats writes.

This tight coupling meant OomAdjuster had knowledge of
ActivityManagerService-specific implementation details.

To improve separation of concerns, this CL introduces a new callback,
`onOomAdjUpdated()`, to the OomAdjuster.Callback interface. The
post-adjustment logic is moved from OomAdjuster to the implementation
of this callback in ActivityManagerService. OomAdjuster is now only
responsible for notifying its client when an adjustment pass is
finished.

Bug: 441408003
Test: m services.core
Test: atest MockingOomAdjusterTests OomAdjusterTests
            ActivityManagerServiceTest
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I407d9daf1c6f32f64ea2940596f12616447e24a5