Update test after flicker visibility computation change.
The visibility computation used by Flicker has been made more rigorous,
checking that a layer's corner radius is less than the layer in front of
it when determining occlusion state instead of only checking occlusion
for layers of equal corner radius.

Accordingly, this test is modified to check that the divider is always
visible or invisible only due to occlusion.

Bug: 312910010
Test: atest com.android.wm.shell.flicker.splitscreen.SwitchAppByDoubleTapDividerTest#splitScreenDividerKeepVisible[ROTATION_0_GESTURAL_NAV]
Flag: EXEMPT TEST_ONLY
Change-Id: I9ab908555d60c8d77fc8e37e74205b7fdc0eeb18