Ensure the task becomes invisible when starting a bubble drag
Long pressed on an expanded bubble will start a bubble drag. The
bubble task used to be moved to the bottom and became invisible
when starting dragging. The task will be moved to top and be
visible again when drag release (without dismissing the bubble).
However, mBubbleData#isExpanded remains true while dragging,
which caused the transitions aborted due to no task visibility
changes.

In addition, the task should also become invisible when switch
to a Chat Bubble or the bubble overflow.

Bug: 407669465
Test: long press on an expanded bubble and release
Test: DismissExpandedBubbleViaBubbleBarTest
Flag: com.android.window.flags.root_task_for_bubble
Change-Id: I1d49a509be9cb7e90998b5b551991aab8715a263