Generate cgroup v2 path instead of looking up
Currently about 4% of the time this puller times out (1.5s). In an
effort to reduce the amount of timeouts, generate the cgroup v2 path for
each process instead of asking the kernel for it.

ActivityManagerService maintains a list of all processes with their UIDs
and PIDs. UIDs are now passed in addition to PIDs for calls to
readMemcgMemorySnapshot and readHighWaterMarkMemorySnapshot so that the
path can be generated internally.

Bug: 439901017
Bug: 446934021
Flag: com.android.server.stats.add_memcg_memory_information_puller
Test: https://data.corp.google.com/android/test?type=atom&atoms=10252
Change-Id: I420a599adb97b626f8e51258f4ba392530db16f2