Update Interactivity on device-state change
When LogicalDisplayMapper (LDM) receives a device-state change,
it can have a stale interactivity value because interactivity is
updated asynchronously via Handler. In rare circumstances, this
can result in a foldable not waking up when unfolded.

Solution is to update the interactivity before we execute a device
state change in LDM.

Bug: 403251804
Test: UnfoldToLockscreenWithNotifications#testUnfold
Flag: EXEMPT BUGFIX
Change-Id: Iaa8548ca8a3fef4fc91fa67ac896be4e6bcb9047