Fix lock on unplug
On display removed is called before on power group removed, this means
that in some situations, the power group a) still exists and b) does not get
updated with the off state of the device.

This fix changes the check to be after the entire power group is
removed. The behaviour should stay the same because:
- If the remainder of the power group is off, then we should automatically
lock anyway.
- If the remainder of the group is on, then there is no reason to lock the
system.
- And if no displays remain, then this code is called.

Bug: 449662568
Flag:com.android.server.power.feature.flags.lock_on_power_group_disconnect
Test: manual

Change-Id: Ic598bb7b94749efbf07281241dcb2801c4e9069c