Allow computeImeParent with null IME Window
Currently if we have no IME window set on the display (e.g. IME process
was killed for some reason), we will also have no IME Layering Target,
making it so computeImeParent returns null. Previously if the IME
Layering Target was null we would still return a non-null value from
computeImeParent, namely the parent of the IME Container (in most cases,
the IME Placeholder), but this was changed with the refactor in [1].

This is normally not an issue, as we will get a new IME process started
automatically, and thus compute a non-null IME parent value. However, if
we enable config_preventImeStartupUnlessTextEditor then we won't have an
IME process until an EditText is focused, and will kill the process as
soon as we lose focus to any EditText. With this config, before we show
the IME the first time, the IME Parent will be null, keeping the IME
Container parented to the IME Placeholder. When the IME is shown, the
IME Parent is updated to the activity that requested it, however as soon
as we exit it, the activity is destroyed, the IME process is killed and
the IME Parent becomes null, but this won't reparent the IME Container,
leaving its surface parented to a destroyed surface. This is also not a
problem in the general case, as we will restart the IME Process whenever
it is requested, recompute the IME Parent to a non-null value, and then
reparent the IME Container.

However, this proves a big problem for the IME Switcher Menu, which is a
child of the IME Container, but can be shown independently of the IME.
If we try to show the menu before showing the IME in any app, this works
well, as the IME Container is still on screen. As soon as we have shown
the IME in some app and then exited it, the IME Switcher Menu cannot be
shown if the IME is not also shown, as the IME Container is left
offscreen.

This fixes the issue by allowing computeImeParent to return the parent
of the IME Container (i.e. the IME Placeholder) if the IME Window is
null. This does not check the values of the IME Layering Target and IME
Input Target as these are also null, or will immediately become null,
when the IME Window is null.

This also fixes another issue in that the IME Switcher Menu could not
be focused with config_preventImeStartupUnlessTextEditor enabled, if
the IME is not currently visible, as the IME Parent would be null, which
breaks some conditions added in [2].

  [1]: I344333a3d0eedf780cd47152e0f4c20022cb5fed
  [2]: I3ac91cf02f0270b2d54581aa9174a875836ca692

Flag: android.view.inputmethod.compute_ime_parent_null_ime_window
Bug: 446170643
Bug: 446170524
Bug: 406501026
Test: atest DisplayContentTests#testComputeImeParent_noImeWindow
Change-Id: I529d3747b4e484d1552c6876315b0e35381bc343