Apply client-drawn corner radii and update opacity on each traversal
Move the logic for applying client-drawn corner radii to a new helper
method, `prepareCornerRadiiForDraw`, and call it in
`performDraw`. This ensures that the corner radii are updated on
every frame, which is necessary when they change dynamically.

This patch also makes the following necessary adjustments for corner
radius handling:

* Makes `ThreadedRenderer#setCornerRadius` return the calculated
    bounding rectangle for the corner radii, which is then passed to
    `setClientDrawnCornerRadius`.
* Updates `ThreadedRenderer#setCornerRadius` to explicitly call
    `setOpaque(false)` when corner radii are set, as a surface with
    rounded corners cannot be fully opaque.

Bug: 438902609
Flag:com.android.graphics.surfaceflinger.flags.set_client_drawn_corner_radii
Test: manual WM smoke test

Change-Id: I2a8fe7ca0622b6da0a7f45a0a8b5409c3438c4e4