Bind to vpn service w/ system uid
When the com.android.server.am.service_check_calling_pkg flag is rolled
out, the system will start enforcing that the calling package matches
the calling uid.

Currently, ConnectivyManager VPN code binds to the VpnService using the
calling UID of the caller but the callingPackage from the system context
which will fail the security check once it's rolled out.

This CL clears the calling identity for the service binding.

Bug: 308138385
Flag: EXEMPT BUGFIX
Test: atest CtsHostsideNetworkTests -- --test-arg com.android.testutils.ConnectivityTestTargetPreparer:ignore-mobile-data-check:true
Change-Id: Id7b9ae196bb251471cf3040b2d0379b6477b6c2a